Use byte for int_to_string(ctr, 1)

+++ core/crypto/draft-irtf-cfrg-vrf-06/ecvrf_p256_sha256_tai_test.go
@@ -44,7 +44,7 @@ func TestVectorsECVRF_P256_SHA256_TAI(t *testing.T) {
 		SK      []byte
 		PK      []byte
 		alpha   []byte
-		wantCtr uint
+		wantCtr byte
 		H       []byte
 		k       []byte
 		U       []byte // k*B
